requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
redis>=5.0.0
asyncpg>=0.29.0
python-binance>=1.0.19
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
websockets>=12.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
filelock>=3.13.0
scikit-learn>=1.4.0
